backend/services/backtesting-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
structlog==23.2.0
python-dotenv==1.0.0

# Database support
asyncpg==0.29.0

# Redis for caching
redis[hiredis]==5.0.1

# HTTP client for service communication
httpx==0.25.2

# JWT authentication
PyJWT==2.8.0

# Data processing
pandas==2.1.3
numpy==1.25.2

# Object storage (optional)
boto3==1.34.0

# Async support
asyncio-throttle==1.0.2